+++ include/graph/storage/SegmentType.h
@@ -13,10 +13,10 @@
 
 namespace graph::storage {
 
-	enum class SegmentType : uint8_t { Node = 0, Edge = 1, Property = 2, Blob = 3 };
+	enum class SegmentType : uint32_t { Node = 0, Edge = 1, Property = 2, Blob = 3 };
 
-	constexpr uint8_t toUnderlying(SegmentType type) {
-		return static_cast<uint8_t>(type);
+	constexpr uint32_t toUnderlying(SegmentType type) {
+		return static_cast<uint32_t>(type);
 	}
 
 } // namespace graph::storage
